Fixed user input check
When the user would type a command, the tests I did in the arduino IDE serial monitor would append a \r to the end of the command, which didn't seem to happen when I tested it in Mission Planner terminal.  I have now written it to trim white space from the submitted command so now it works in both the IDE serial monitor and Mission Planner terminal. Command input works in both now with expected output.